requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.6
fiona==1.9.5
fonttools==4.46.0
frozenlist==1.4.0
geopandas==0.14.1
idna==3.6
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
multidict==6.0.4
networkx==3.2.1
numpy==1.26.2
osmnx==1.8.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pydantic==2.5.2
pyparsing==3.1.1
pyproj==3.6.1
pyrobuf==0.9.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
shapely==2.0.2
six==1.16.0
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.4
